backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.3
python-multipart>=0.0.6

# AI & Agent
ollama>=0.1.5
# We use a custom robust agent implementation, but keep these for future expansion
langchain>=0.1.0
langchain-community>=0.0.13
langchain-core>=0.1.10

# Web & Tools
duckduckgo-search>=4.1.1
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.1.0

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0

# Voice
edge-tts>=6.1.9

# Database & Utils
aiosqlite>=0.19.0
python-dateutil>=2.8.2

# Optional ML (RAG) - These are heavy, install if possible
# sentence-transformers>=2.3.1
# chromadb>=0.4.22
